requirements.txt
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
Flask==3.0.1
Flask-Session==0.6.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
Werkzeug==3.0.1
